+++ SandboxiePlus/QSbieAPI/Sandboxie/SandBox.h
@@ -59,11 +59,15 @@ public:
 	virtual bool					IsEnabled() const  { return m_IsEnabled; }
 
 	virtual bool					IsEmpty() const;
+	virtual bool					IsInitialized() const;
+	virtual bool					HasSnapshots() const;
 	virtual SB_PROGRESS				CleanBox();
 	virtual SB_STATUS				RenameBox(const QString& NewName);
 	virtual SB_STATUS				RemoveBox();
 
-	virtual QList<SBoxSnapshot>		GetSnapshots(QString* pCurrent = NULL) const;
+	virtual QList<SBoxSnapshot>		GetSnapshots(QString* pCurrent = NULL, QString* pDefault = NULL) const;
+	virtual void					SetDefaultSnapshot(QString Default);
+	virtual QString					GetDefaultSnapshot(QString* pCurrent = NULL) const;
 	virtual SB_PROGRESS				TakeSnapshot(const QString& Name);
 	virtual SB_PROGRESS				RemoveSnapshot(const QString& ID);
 	virtual SB_PROGRESS				SelectSnapshot(const QString& ID);