+++ SandboxiePlus/SandMan/Windows/RecoveryWindow.cpp
@@ -49,6 +49,7 @@ CRecoveryWindow::CRecoveryWindow(const CSandBoxPtr& pBox, QWidget *parent)
 	m_LastTargetIndex = 0;
 	m_bTargetsChanged = false;
 	m_bReloadPending = false;
+	m_DeleteShapshots = false;
 
 #ifdef WIN32
 	QStyle* pStyle = QStyleFactory::create("windows");
@@ -99,6 +100,10 @@ CRecoveryWindow::CRecoveryWindow(const CSandBoxPtr& pBox, QWidget *parent)
 	ui.btnRecover->setPopupMode(QToolButton::MenuButtonPopup);
 	ui.btnRecover->setMenu(pRecMenu);
 
+	QMenu* pDelMenu = new QMenu(ui.btnDeleteAll);
+	pDelMenu->addAction(tr("Delete evertyhign including all snapshots"), this, SLOT(OnDeleteEverything()));
+	ui.btnDeleteAll->setPopupMode(QToolButton::MenuButtonPopup);
+	ui.btnDeleteAll->setMenu(pDelMenu);
 
 	restoreGeometry(theConf->GetBlob("RecoveryWindow/Window_Geometry"));
 
@@ -145,6 +150,7 @@ int	CRecoveryWindow::exec()
 {
 	//QDialog::setWindowModality(Qt::WindowModal);
 	ui.btnDeleteAll->setVisible(true);
+	SafeShow(this);
 	return QDialog::exec();
 }
 
@@ -221,6 +227,12 @@ void CRecoveryWindow::OnDeleteAll()
 {
 	this->setResult(1);
 	this->close();
+}
+
+void CRecoveryWindow::OnDeleteEverything()
+{
+	m_DeleteShapshots = true;
+	OnDeleteAll();
 }
 
 void CRecoveryWindow::AddFile(const QString& FilePath, const QString& BoxPath)